requirements.txt
tensorflow==2.10.0
scikit-learn==1.1.2
Pillow==9.4.0
numpy==1.23.4
opencv-python==4.7.0.72
matplotlib==3.5.2
pyyaml